+++ src/main/java/tariffs/FixedTariff.java
@@ -2,16 +2,16 @@ package tariffs;
 
 public class FixedTariff extends Tariff{
 
-	public FixedTariff(int pricePerKW, int energyCap) {
+	public FixedTariff(float pricePerKW, float energyCap) {
 		super(pricePerKW, energyCap);
 	}
 	
-	public int calculateOffer(int energyWanted) {
+	public float calculateOffer(float energyWanted) {
 		energyWanted = energyWanted > this.energyCap ? this.energyCap : energyWanted;
 		return this.calculatePrice(5, energyWanted);
 	}
 	
-	public int calculatePrice(int penaltyRate, int energyWanted) {
+	public float calculatePrice(float penaltyRate, float energyWanted) {
 		return this.pricePerKW * penaltyRate * energyWanted;
 	}
