+++ src/main/java/tariffs/TariffInterface.java
@@ -2,8 +2,8 @@ package tariffs;
 
 public interface TariffInterface {
 
-	int calculateOffer(int energyWanted);
-	int calculatePrice(int penaltyRate, int energyWanted);
-	int getCap();
+	float calculateOffer(float energyWanted);
+	float calculatePrice(float penaltyRate, float energyWanted);
+	float getCap();
 	
 }
